Fix wrangler test process helpers
